requirements.txt
streamlit
pandas
numpy
scikit-learn
openai
python-dotenv
plotly
tqdm
fastapi
uvicorn
redis
requests
